+++ frontend/webadmin/modules/gwt-common/src/main/java/org/ovirt/engine/ui/common/widget/SliderBar.java
@@ -813,8 +813,8 @@ public class SliderBar extends FocusPanel implements RequiresResize,
         Element knobElement = knobImage.getElement();
         int lineWidth = lineElement.getOffsetWidth();
         int knobWidth = knobElement.getOffsetWidth();
-        int knobLeftOffset = (int) (lineLeftOffset
-                + ((int) getKnobPercent() * lineWidth) - (knobWidth / 2));
+        int knobLeftOffset = lineLeftOffset
+                + ((int) getKnobPercent() * lineWidth) - (knobWidth / 2);
         knobLeftOffset = Math.min(knobLeftOffset, lineLeftOffset + lineWidth
                 - (knobWidth / 2) - 1);
         DOM.setStyleAttribute(knobElement, "left", knobLeftOffset + "px"); //$NON-NLS-1$ //$NON-NLS-2$
